review changes : use mui components and copy info not truncated
