--- src/components/CopyToClipboard.tsx
@@ -1,44 +1,0 @@
-interface props {
-  textToCopy: string;
-}
-
-export const CopyToClipboard = (props: props) => {
-  const { textToCopy } = props;
-
-  async function handleCopy() {
-    try {
-      await navigator.clipboard.writeText(textToCopy);
-    } catch (error) {
-      console.error('Failed to copy to clipboard:', error);
-    }
-  }
-  return (
-    <div
-      onClick={() => void handleCopy()}
-      style={{
-        cursor: 'pointer',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <rect width="8" height="4" x="8" y="2" rx="1" ry="1" />
-        <path d="M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" />
-        <path d="M16 4h2a2 2 0 0 1 2 2v4" />
-        <path d="M21 14H11" />
-        <path d="m15 10-4 4 4 4" />
-      </svg>
-    </div>
-  );
-};
